Reorganizing code, removing stale simple_travel_example.h file
