+++ apps/planners/domains/simple_travel.h
@@ -1,4 +1,4 @@
-#include "../cpphop.h"
+#include "cpphop.h"
 
 auto taxi_rate(double dist) { return 1.5 + 0.5 * dist; }
