+++ src/cpp/cpphop.h
