+++ src/cpp/plan_trace.h
